pyproject.toml
[project]
name = "handsoff-poetry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.74.0.post1",
    "openai>=1.93.3",
    "openai-agents>=0.1.0",
    "py-dotenv>=0.1",
]
